TST/REF: misplaced Series.reindex test
