+++ pandas/tests/series/methods/test_reindex.py
@@ -1,8 +1,16 @@
 import numpy as np
 import pytest
 
-import pandas as pd
-from pandas import Categorical, Series, date_range, isna
+from pandas import (
+    Categorical,
+    Index,
+    NaT,
+    Period,
+    PeriodIndex,
+    Series,
+    date_range,
+    isna,
+)
 import pandas._testing as tm
 
 
@@ -293,5 +301,33 @@ def test_reindex_datetimeindexes_tz_naive_and_aware():
 def test_reindex_empty_series_tz_dtype():
     # GH 20869
     result = Series(dtype="datetime64[ns, UTC]").reindex([0, 1])
-    expected = Series([pd.NaT] * 2, dtype="datetime64[ns, UTC]")
+    expected = Series([NaT] * 2, dtype="datetime64[ns, UTC]")
     tm.assert_equal(result, expected)
+
+
+@pytest.mark.parametrize(
+    "p_values, o_values, values, expected_values",
+    [
+        (
+            [Period("2019Q1", "Q-DEC"), Period("2019Q2", "Q-DEC")],
+            [Period("2019Q1", "Q-DEC"), Period("2019Q2", "Q-DEC"), "All"],
+            [1.0, 1.0],
+            [1.0, 1.0, np.nan],
+        ),
+        (
+            [Period("2019Q1", "Q-DEC"), Period("2019Q2", "Q-DEC")],
+            [Period("2019Q1", "Q-DEC"), Period("2019Q2", "Q-DEC")],
+            [1.0, 1.0],
+            [1.0, 1.0],
+        ),
+    ],
+)
+def test_reindex_periodindex_with_object(p_values, o_values, values, expected_values):
+    # GH#28337
+    period_index = PeriodIndex(p_values)
+    object_index = Index(o_values)
+
+    ser = Series(values, index=period_index)
+    result = ser.reindex(object_index)
+    expected = Series(expected_values, index=object_index)
+    tm.assert_series_equal(result, expected)
